core code updates
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@5125 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/Scanner.java
@@ -2181,7 +2181,7 @@ public class Scanner {
         int    firstPart     = type.startPartIndex;
         int    lastPart      = type.endPartIndex;
         long   totalValue    = 0;
-        int    currentValue  = 0;
+        long   currentValue  = 0;
         int    i             = intervalPosition;
         int    currentPart   = firstPart;
         int    currentDigits = 0;
@@ -2232,7 +2232,7 @@ public class Scanner {
 
                     int factor = factors[currentPart];
 
-                    totalValue        += (long) currentValue * factor;
+                    totalValue        += currentValue * factor;
                     intervalPrecision = currentDigits;
                 } else {
                     if (currentValue >= limits[currentPart]) {
